refactor(webapp): rename Producers components
